apps/02-deep-research/requirements.txt
llama-stack-client
tavily-python
deepagents

